+++ rtdb-limit-child-nodes/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Version 0.1.18
+
+chore: bump dependencies
+
 ## Version 0.1.17
 
 chore: update Cloud Functions runtime to Node.js 22